+++ src/main/java/dev/fulmineo/companion_bats/item/CompanionBatItem.java
@@ -42,8 +42,9 @@ public class CompanionBatItem extends Item {
             CompoundTag entityData = getOrCreateEntityData(itemStack);
             if (user.isSneaking()){
                 SimpleInventory inventory = new SimpleInventory(3);
-                inventory.setStack(0, ItemStack.fromNbt(entityData.getCompound("armor")));
-				inventory.setStack(1, ItemStack.fromNbt(entityData.getCompound("bundle")));
+                inventory.setStack(0, ItemStack.fromNbt(entityData.getCompound("accessory")));
+                inventory.setStack(1, ItemStack.fromNbt(entityData.getCompound("armor")));
+				inventory.setStack(2, ItemStack.fromNbt(entityData.getCompound("bundle")));
                 user.openHandledScreen(new ExtendedScreenHandlerFactory() {
                     @Override
                     public void writeScreenOpeningData(ServerPlayerEntity serverPlayerEntity, PacketByteBuf packetByteBuf) {